Some finish modifies

+++ src/main.cpp
@@ -41,7 +41,7 @@ int main(int argc, char **argv) {
 		}
 	} else if (parser.isSet("c")) {
 		if (parser.isSet("p")) {
-			Kyrys::Client client("localhost", parser.value("p").toInt());
+			Kyrys::Client client("localhost", parser.value("p").toUShort());
 			return a.exec();
 		} else {
 			Kyrys::Client client("localhost", defaultPort);
